move search stream types to own files
